TST: Added common test functionality for VTK format.
