+++ src/porepy/applications/test_utils/__init__.py
@@ -4,5 +4,4 @@ Utility functions that are not specific to a single test module should be placed
 Functions which are relevant also outside tests should go elsewhere.
 
 """
-from . import arrays
-from . import grids
+from . import arrays, grids, vtk
