--- visualization/times.json
@@ -1,1 +1,0 @@
-{"time": [0, 1], "dt": [1.0, 1.0]}
